+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Legenda for the issue kind:
 - [#539](https://github.com/ariadne-cps/ariadne/issues/539) (F) A segmentation fault sometimes would be issued when terminating the executable, due to logging
 - [#557](https://github.com/ariadne-cps/ariadne/issues/557) (F) Fix behavior of StopWatch utility for concurrent code, enhance the class for choosing a duration type
 - [#567](https://github.com/ariadne-cps/ariadne/issues/567) (F) Fix problem with state_time space creation when state space already contains the 't' variable
+- [#576](https://github.com/ariadne-cps/ariadne/issues/576) (F) Fix HybridEnclosure state_set() incorrectly working when an auxiliary function is present
 - [#558](https://github.com/ariadne-cps/ariadne/issues/558) (R) Remove unnecessary IteratedMapEvolver::enclosure methods
 - [#552](https://github.com/ariadne-cps/ariadne/issues/552) (R) Remove evolve/reach/reach_evolve methods from EvolverInterface, relying on orbit generation only
 - [#553](https://github.com/ariadne-cps/ariadne/issues/553) (R) Remove ability to write an evolver object to the standard output, since it was implemented as a fixed string for all evolvers anyway